+++ functions/Get-TibberUser.ps1
@@ -20,6 +20,7 @@
             'userId'
             'name'
             'accountType'
+            'websocketSubscriptionUrl'
         )
     )